Add FrequencyVisualizer service for live dual-band frequency data
Introduce a FrequencyVisualizer that encapsulates dual-band FFT analysis
(lowpass + highpass biquad filters with separate AnalyserNodes) and
log-frequency mapping. All consumers (Mixer, MicrophoneUserMedia,
plasmaRenderer, RecordingBuffer, Spectrogram) now receive pre-processed
Uint8Array data — no frequency mapping details leak outside the service.

+++ src/components/workstation/scrubber/PlasmaPlayhead.tsx
@@ -10,7 +10,7 @@ export const PLASMA_WIDTH = 240;
 
 export type PlasmaPlayheadHandle = {
   render: (
-    frequencyData: Float32Array | null,
+    frequencyData: Uint8Array | null,
     loudness: number,
     scrollLeft: number,
     trackFrequencyInputs: TrackFrequencyInput[],
@@ -33,7 +33,7 @@ const PlasmaPlayhead = forwardRef<PlasmaPlayheadHandle, PlasmaPlayheadProps>(
 
     useImperativeHandle(ref, () => ({
       render(
-        frequencyData: Float32Array | null,
+        frequencyData: Uint8Array | null,
         loudness: number,
         scrollLeft: number,
         trackFrequencyInputs: TrackFrequencyInput[],
